Fix issue #19:L mlc fails with "In cpu range field 32 > 0".
The input file mlc_loaded_latency.input file sets the range of CPU's used to measue the bandwidth.  In this case, we need to make sure that the range is correct.  FIRST_CPU_ON_SOCKET - TO_CPU.

In addition to the fix for the range, I have added code to call get_first_cpu_in_socket in the functions that use FIRST_CPU_ON_SOCKET as a safety measure.